requirements.txt
Django
markdown
django-annoying
django-filter
djangorestframework
pip
psycopg2
requests
setuptools
wsgiref
